+++ MultiEducation/ViewModels/MEVM.h
@@ -81,6 +81,6 @@
  @param success callback
  @param failure callback
  */
-- (void)postData:(NSData *_Nonnull)data pageSize:(int32_t)size pageIndex:(int32_t)index hudEnable:(BOOL)hud success:(void (^_Nullable)(NSData * _Nullable resObj, NSUInteger totalPages))success failure:(void (^_Nullable)(NSError * _Nonnull error))failure;
+- (void)postData:(NSData *_Nonnull)data pageSize:(int32_t)size pageIndex:(int32_t)index hudEnable:(BOOL)hud success:(void (^_Nullable)(NSData * _Nullable resObj, int32_t totalPages))success failure:(void (^_Nullable)(NSError * _Nonnull error))failure;
 
 @end
